third_party/blink/web_tests/external/wpt/css/css-sizing/animation/max-width-interpolation-expected.txt

This is a testharness.js-based test.
Found 42 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-behavior:allow-discrete: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition-property:all and transition-behavor:allow-discrete: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (-0.3) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (0) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (0.3) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (-0.3) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.3) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] CSS Animations: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (-0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (0) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (0.3) should be [stretch]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (0.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (0.6) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (1) should be [none]
  assert_true: 'from' value should be supported expected true got false
[FAIL] Web Animations: property <max-width> from [stretch] to [none] at (1.5) should be [none]
  assert_true: 'from' value should be supported expected true got false
Harness: the test ran to completion.
